refactor: add connect wallet on modal
