+++ .gitignore
@@ -19,5 +19,5 @@
 .env.production.local
 
 npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
+
+
