+++ src/componentes/Navbar/Navbar.jsx
@@ -24,7 +24,8 @@ import {
 import { useEffect, useState } from "react";
 import { SwitchNetwork, connectWallet } from "../Button/utils.jsx";
 import { Modal } from "../Modals/Modal";
-import { PortalRoot } from "../PortalRoot/PortalRoot";
+import { PortalRoot } from "../PortalModal/PortalRoot";
+import { PortalSwiper } from "../PortalModal/PortalSwiper";
 
 const Navbar = () => {
   const [selectedNetwork, setSelectedNetwork] = useState(networkData[0]);
@@ -108,7 +109,9 @@ const Navbar = () => {
         />
         {isModalVisible && (
           <Modal setIsModalVisible={setIsModalVisible}>
-            <PortalRoot connectWallet={connectWallet(setWallet)}/>
+            <PortalRoot
+              connectWallet={() => connectWallet(setWallet, selectedNetwork)}
+            />
           </Modal>
         )}
       </div>
