+++ src/componentes/PortalModal/PortalRoot.scss
@@ -68,29 +68,40 @@
 
 .ventana-prime {
   display: flex;
+
   width: 100%;
   height: 100%;
+
   position: absolute;
+
   background: var(--colors-gradientCardHeader);
+
   border-radius: 32px;
   -webkit-border-radius: 32px;
   -moz-border-radius: 32px;
   -ms-border-radius: 32px;
   -o-border-radius: 32px;
+
   > .ventana-primaria {
     display: flex;
     flex-flow: row wrap;
     text-align: justify;
     align-content: space-around;
+
     box-sizing: border-box;
+
     background-color: #27262c;
+
     padding: 32px 0px;
+
     width: 53%;
+
     border-radius: 32px;
     -webkit-border-radius: 32px;
     -moz-border-radius: 32px;
     -ms-border-radius: 32px;
     -o-border-radius: 32px;
+
     > h2 {
       padding: 0px 48px;
       width: 100%;
@@ -115,10 +126,6 @@
     width: 100%;
 
     margin: 0px 24px;
-
-    > div {
-      display: flex;
-    }
 
     > img {
       flex: 1 1;
